Catch errors from audioElement.play()
HTMLMediaElement.play() returns a promise which is resolved when
playback has been started, or is rejected if for any reason playback
cannot be started.

Possible errors include:

NotAllowedError: The user agent (browser) or operating system doesn't
allow playback of media in the current context or situation. This may
happen, for example, if the browser requires the user to explicitly
start media playback by clicking a "play" button.

NotSupportedError: The media source (which may be specified as a
MediaStream, MediaSource, Blob, or File, for example) doesn't represent
a supported media format.